requirements.txt
certifi==2020.12.5
colorama==0.4.4
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
qrcode==6.1
six==1.15.0
wincertstore==0.2
